Migrate more sensors to Wear OS
