+++ common/src/main/java/io/homeassistant/companion/android/common/sensors/DisplaySensorManager.kt
@@ -1,9 +1,8 @@
-package io.homeassistant.companion.android.sensors
+package io.homeassistant.companion.android.common.sensors
 
 import android.content.Context
 import android.provider.Settings
 import android.util.Log
-import io.homeassistant.companion.android.common.sensors.SensorManager
 import io.homeassistant.companion.android.common.R as commonR
 
 class DisplaySensorManager : SensorManager {
