+++ common/src/main/java/io/homeassistant/companion/android/common/sensors/StorageSensorManager.kt
@@ -1,10 +1,9 @@
-package io.homeassistant.companion.android.sensors
+package io.homeassistant.companion.android.common.sensors
 
 import android.content.Context
 import android.os.Environment
 import android.os.StatFs
 import android.util.Log
-import io.homeassistant.companion.android.common.sensors.SensorManager
 import java.io.File
 import kotlin.math.roundToInt
 import io.homeassistant.companion.android.common.R as commonR
