+++ common/src/main/java/io/homeassistant/companion/android/common/sensors/TrafficStatsManager.kt
@@ -1,4 +1,4 @@
-package io.homeassistant.companion.android.sensors
+package io.homeassistant.companion.android.common.sensors
 
 import android.content.Context
 import android.net.ConnectivityManager
@@ -6,7 +6,6 @@ import android.net.NetworkCapabilities
 import android.net.TrafficStats
 import android.util.Log
 import androidx.core.content.getSystemService
-import io.homeassistant.companion.android.common.sensors.SensorManager
 import java.math.RoundingMode
 import kotlin.math.absoluteValue
 import io.homeassistant.companion.android.common.R as commonR
